--add test function and paths to fix unit tests.  Replace sprintf with fmt calls, faster and more secure. Replace atoi/atoi insecure C calls with C++ std::stoi, etc. update semgrep.sh script for security scans. Create RAPIORuntime for generic library initialization.

+++ modules/iohmrg/rHmrgProductInfo.cc
@@ -45,9 +45,9 @@ ProductInfoSet::readConfigFile()
           line >> varPrefix;
           line >> productDir;
           line >> temp;
-          varMissing = (temp == "none") ? ProductInfo::UNDEFINED : atof(temp.c_str());
+          varMissing = (temp == "none") ? ProductInfo::UNDEFINED : std::stof(temp.c_str());
           line >> temp;
-          varNoCoverage = (temp == "none") ? ProductInfo::UNDEFINED : atof(temp.c_str());
+          varNoCoverage = (temp == "none") ? ProductInfo::UNDEFINED : std::stof(temp.c_str());
           line >> varScale;
           line >> w2Name;
           line >> w2Unit;
